Lager ApplicationContext klasse innsendt-soeknad
